criação de exceção para banco de dados, alteração da transação do metodo delete para capturar e dar rethrow no proprio service

+++ dscommerce/src/main/java/com/devsuperior/felipe/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.devsuperior.felipe.dscommerce.services;
 import com.devsuperior.felipe.dscommerce.dto.ProductDTO;
 import com.devsuperior.felipe.dscommerce.entities.Product;
 import com.devsuperior.felipe.dscommerce.repositories.ProductRepository;
+import com.devsuperior.felipe.dscommerce.services.exceptions.DatabaseException;
 import com.devsuperior.felipe.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -44,15 +48,28 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        this.copyDTOtoEntity(dto,entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            //if (entity == null) throw new ResourceNotFoundException("Produto não encontrado"); poderia fazer assim também
+            this.copyDTOtoEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException ex) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private Product fromDTO(ProductDTO dto) {
